Renaming inner directory
