+++ numanalysis/approximations/discrete.py
