+++ numanalysis/approximations/funsys.py
