+++ numanalysis/approximations/quadraticapproximation.py
