+++ numanalysis/approximations/spline.py
