+++ numanalysis/approximations/util.py
