+++ numanalysis/integral/integration_strategy.py
