+++ numanalysis/nonlineareq/non_linear_equations.py
